refactor: split remote worker reporter
